requirements.txt
GitPython
jsonschema
pandas
tqdm
pylint
pytest
pytest-dependency
pytest-order
requests